+++ src/libtrx/include/libtrx/game/demo.h
@@ -2,12 +2,12 @@
 
 #include "./game_flow/types.h"
 
-bool Demo_Start(int32_t level_num);
-void Demo_End(void);
-void Demo_Pause(void);
-void Demo_Unpause(void);
-void Demo_StopFlashing(void);
+extern bool Demo_Start(int32_t level_num);
+extern void Demo_End(void);
+extern void Demo_Pause(void);
+extern void Demo_Unpause(void);
+extern void Demo_StopFlashing(void);
 
-bool Demo_GetInput(void);
-GF_COMMAND Demo_Control(void);
-int32_t Demo_ChooseLevel(int32_t demo_num);
+extern bool Demo_GetInput(void);
+extern GF_COMMAND Demo_Control(void);
+extern int32_t Demo_ChooseLevel(int32_t demo_num);