+++ docs/tr2/CHANGELOG.md
@@ -9,6 +9,7 @@
 - added a fade-out effect when completing Lara's Home
 - added support for animated sprites (#2401)
 - added a `/cut` console command for playing cutscenes
+- added a console log when using the `/demo` command
 - changed the object texture limit from 2048 to unlimited (within game's overall memory cap) (#1795)
 - changed the sprite texture limit from 512 to unlimited (within game's overall memory cap) (#1795)
 - changed the texture page limit from 32 to 128 (#1796)